% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logis}
\alias{logis}
\title{logistic function}
\usage{
logis(x, max = 1, k = 1, x_0 = 0)
}
\arguments{
\item{x}{the value to transform}

\item{max}{the assymptote of the function}

\item{k}{the rate of the function}

\item{x_0}{the inflection point}
}
\value{
returns the transformed value of the input
}
\description{
logistic function
}
