% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{make_matrix}
\alias{make_matrix}
\title{A function to create an M matrix to help with
restricted cubic spline calculations}
\usage{
make_matrix(
  precision,
  k,
  dimensions = 1L,
  min = 0,
  max = 1,
  first_knot = 0.05,
  last_knot = 0.95
)
}
\arguments{
\item{precision}{how precise the calculation should be;
it will the be the number of rows in the output}

\item{k}{the number of knots used}

\item{dimensions}{the number of dimentions to create
(this grows the size VERY fast)}

\item{min}{the minimum value of the x values; default = 0}

\item{max}{the maximum value of the x values; default = 1}

\item{first_knot}{position of the first knot; default = 0.05 * max}

\item{last_knot}{position of the last knot; default = 0.95 * max}
}
\value{
a matrix of size precision by k
}
\description{
A function to create an M matrix to help with
restricted cubic spline calculations
}
\examples{
make_matrix(1000, 5)
}
