% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spline_1d}
\alias{spline_1d}
\title{A function that takes a vector and compatable matrix and
returns a restricted cubic spline}
\usage{
spline_1d(vector, matrix)
}
\arguments{
\item{vector}{A vector}

\item{matrix}{The M matrix to calculate the restricted cubic spline;
number of columns should be the same as the length of the vector}
}
\value{
Vector with the spline values; length is the same as the number
of rows in matrix
}
\description{
A function that takes a vector and compatable matrix and
returns a restricted cubic spline
}
\examples{
M <- make_matrix(1000, 5)
spline_1d(v, M)
}
